zipkin_c: in zipkin, port is a (signed!) 16-bit integer
The Zipkin thrift structure definitions explicitly chose a signed
representation, and advice callers to convert in advance.

Yes, the blkin encoders throw when out-of-range (high-numbered) port
values are seen.

+++ blkin-lib/zipkin_c.c
@@ -135,8 +135,8 @@ OUT:
 	return res;
 }
 
-int blkin_init_endpoint(struct blkin_endpoint *endp, const char *ip, int port,
-			const char *name)
+int blkin_init_endpoint(struct blkin_endpoint *endp, const char *ip,
+			int16_t port, const char *name)
 {
 	int res;
 	if (!endp){
